+++ src/components/anagrammGame/components/Randomizer.js
@@ -1,0 +1,4 @@
+export default function Randomizer(number) {
+    const randomIndex = Math.floor(Math.random() * (number + 1));
+    return randomIndex;
+}
