--- src/components/anagrammGame/components/StartPage.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import ShuffleLettersBox from './ShuffleLettersBox';
-
-export default function Wrapper() {
-    return (
-        <div>
-            <ShuffleLettersBox />
-        </div>
-    );
-}
